revert test

+++ paddle/fluid/pir/transforms/onednn/elementwise_act_onednn_fuse_pass.cc
@@ -83,7 +83,7 @@ class ElementwiseActivationFusePattern : public paddle::drr::DrrPatternBase {
 
     pat.AddConstraint([&](const paddle::drr::MatchContext &match_ctx) {
       if (activation_name_ == "leaky_relu") {
-        double negative_slope = match_ctx.Attr<double>("negative_slope");
+        float negative_slope = match_ctx.Attr<float>("negative_slope");
         // leaky relu alpha is a positive number
         if (negative_slope <= 0.0) {
           return false;
